move command to its own package
