+++ order-api/src/apiTest/java/com/ecommerce/order/order/OrderApiTest.java
@@ -1,11 +1,11 @@
 package com.ecommerce.order.order;
 
 import com.ecommerce.order.BaseApiTest;
-import com.ecommerce.order.command.ChangeAddressDetailCommand;
-import com.ecommerce.order.command.ChangeProductCountCommand;
-import com.ecommerce.order.command.CreateOrderCommand;
-import com.ecommerce.order.command.OrderItemCommand;
-import com.ecommerce.order.command.PayOrderCommand;
+import com.ecommerce.order.command.order.ChangeAddressDetailCommand;
+import com.ecommerce.order.command.order.ChangeProductCountCommand;
+import com.ecommerce.order.command.order.CreateOrderCommand;
+import com.ecommerce.order.command.order.OrderItemCommand;
+import com.ecommerce.order.command.order.PayOrderCommand;
 import com.ecommerce.order.order.model.Order;
 import com.ecommerce.order.order.model.OrderId;
 import com.ecommerce.order.order.model.ProductId;
