+++ order-api/src/main/java/com/ecommerce/order/order/OrderController.java
@@ -1,9 +1,9 @@
 package com.ecommerce.order.order;
 
-import com.ecommerce.order.command.ChangeAddressDetailCommand;
-import com.ecommerce.order.command.ChangeProductCountCommand;
-import com.ecommerce.order.command.CreateOrderCommand;
-import com.ecommerce.order.command.PayOrderCommand;
+import com.ecommerce.order.command.order.ChangeAddressDetailCommand;
+import com.ecommerce.order.command.order.ChangeProductCountCommand;
+import com.ecommerce.order.command.order.CreateOrderCommand;
+import com.ecommerce.order.command.order.PayOrderCommand;
 import com.ecommerce.order.order.model.OrderId;
 import com.ecommerce.order.order.representation.OrderRepresentationService;
 import com.ecommerce.order.order.representation.detail.OrderRepresentation;
